Remove "Remove form" link from forms Add a different view for all controllers instead of directly using standardForm (now it includes standardForm)
This is because of a problem of update in forms when clearing/removing :
Case 1
In advanced form : Add a Radiative form, add an Environment form
Then Clear function in Environment does not work

In advanced form : Add a Particle form, add an Atom form
Then Clear function in Atom does not work

If a form of the same type is added above another one, Clear does not work (because the ID of the new form is identical to the form below)

Case 2 :
In advanced form : Add 3 atom forms
Enter any value in Atom Symbol in each one (1, 2, 3)
Remove the first form
Then the values in the remaining forms are incorrect

To solve Case 1,  several views were added so that generated objects are differents in the page.
To solve Case 2, "Remove form" has been removed. A form can be hidden and a new function "Clear page" has been added in Advanced query